# backend/openapi/horse-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Horse Reservation API",
    "description": "마장 예약 서비스 API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/admin/jobs/sync-schedule-occurrences/retry": {
      "post": {
        "tags": [
          "admin-schedule-synchronization-controller"
        ],
        "operationId": "retry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleSynchronizationRetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleSynchronizationResultResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/schedule-sync": {
      "get": {
        "tags": [
          "admin-schedule-synchronization-controller"
        ],
        "operationId": "getStatus",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleSynchronizationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ScheduleSynchronizationResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SYNCING"
            ]
          },
          "activeVersion": {
            "type": "integer",
            "format": "int64"
          },
          "pendingVersion": {
            "type": "integer",
            "format": "int64"
          },
          "horizonStart": {
            "type": "string",
            "format": "date"
          },
          "horizonEnd": {
            "type": "string",
            "format": "date"
          },
          "totalDateCount": {
            "type": "integer",
            "format": "int64"
          },
          "appliedDateCount": {
            "type": "integer",
            "format": "int64"
          },
          "remainingDateCount": {
            "type": "integer",
            "format": "int64"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "syncStartedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastCompletedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastFailedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastFailureCode": {
            "type": "string"
          },
          "lastFailureSummary": {
            "type": "string"
          },
          "longRunning": {
            "type": "boolean"
          }
        },
        "required": [
          "activeVersion",
          "appliedDateCount",
          "horizonEnd",
          "horizonStart",
          "longRunning",
          "progressPercent",
          "remainingDateCount",
          "status",
          "totalDateCount"
        ]
      },
...
      "ScheduleSynchronizationRetryRequest": {
        "type": "object",
        "properties": {
          "pendingVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "pendingVersion"
        ]
      },
      "ScheduleSynchronizationResultResponse": {
        "type": "object",
        "properties": {
          "targetVersion": {
            "type": "integer",
            "format": "int64"
          },
          "createdCount": {
            "type": "integer",
            "format": "int32"
          },
          "updatedCount": {
            "type": "integer",
            "format": "int32"
          },
          "appliedDateCount": {
            "type": "integer",
            "format": "int32"
          },
          "skippedDateCount": {
            "type": "integer",
            "format": "int32"
          },
          "synchronization": {
            "$ref": "#/components/schemas/ScheduleSynchronizationResponse"
          }
        },
        "required": [
          "appliedDateCount",
          "createdCount",
          "skippedDateCount",
          "synchronization",
          "targetVersion",
          "updatedCount"
        ]
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "details",
          "fieldErrors",
          "message",
          "path",
          "status",
          "timestamp"
        ]
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}